Made one of the Extra's Features work.
Added the color system. It updates without hitting a button. I'm certain
there's a better way to go about doing it.

I'd like to implement it to where you can click on the colors in the
dweller's menu, and it converts from hexadecimal internally, similar in
system to how lunchboxes are handled.